Backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.3.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.12.0
